* The Multiple End Entry Test
Figure 1 shows an example input file. Like all (correct) input files, it starts with a single header
record, followed by a sequence of text records, and finishes with a single end record.

The above was taken from lab1 specs.
E3001
E3002			! *This line does not causes a fatal error, which is not the desired result.